Fixed persistence of previously completed or newly completed files.
The newer method actually appends to the file instead of overriding it. We persist an in-memory version of every comment we've previously addressed. On top of that, each loop that we run appends every comment with square brackets in it to both the in-memory list and the lolfetcher_done.txt file, before getting wiped at the start of the next run.